Add layout constants and apply across screens
