+++ theme.js
@@ -1,20 +1,3 @@
-import { Platform } from 'react-native';
-
-export const FONT_SIZES = {
-  title: 24,
-  label: 16,
-  text: 14,
-};
-
-export const BUTTON_STYLE = {
-  paddingVertical: 12,
-  paddingHorizontal: 24,
-  borderRadius: 8,
-};
-
-export const HEADER_HEIGHT = 60;
-export const HEADER_SPACING = Platform.OS === 'ios' ? HEADER_HEIGHT + 40 : HEADER_HEIGHT + 20;
-
 export const lightTheme = {
   background: '#ffffff',
   card: '#ffffff',
